refactor: rename sandbox -> extraction; stop asserting isolation
The module named sandbox.py performs no sandboxing: it is a code extractor
(CodeExtractor) and a dependency copier (DependencyMapper). The actual execution
(executor.py) is a temp dir plus a subprocess plus a wall-clock timeout, with no
filesystem jail, network cut-off, resource caps, or syscall filtering, and the
subprocess inherits the invoking user's privileges. The name and the docstrings
asserted a safety property the code does not provide, which the independent
audit (and any reasonable committee) flags as a misnomer.

- Renamed src/qallm/verification/sandbox.py to extraction.py (git rename) and
  updated all seven importers.
- Rewrote the module docstring and the executor docstrings to state the truth:
  not isolated, bounded only by a timeout, run in the provided container and
  never against untrusted notebooks on a host with secrets.
- Added an execution-safety threats-to-validity paragraph to the thesis scaffold
  (5.5) and fixed the stale "under the sandbox" wording in the security
  paragraph. Stronger isolation (network-disabled, resource-capped, read-only
  per-run container) is recorded as low-effort future hardening.

No behaviour change; pure naming and documentation honesty. Suite 718 passed;
ruff clean.
